feat(exec)!: make WebSocket the only exec transport
Drop the opt-in `transport` flag and remove the backboard GraphQL exec
route (sandboxExec mutation, sandboxExecOutput subscription,
sandboxExecKill). exec() now always runs over the /ws/exec bridge.

- merge ws-exec into exec.ts; delete ws-exec.ts and graphql-ws-client.ts
- remove ExecOptions.transport and ExecOptions.waitMs (WS streams from
  the first byte; the inline-drain window is gone)
- ExecHandle.kill() drops the signal arg (WS kill closes the session)
- remove SandboxExecInterruptedError (only the GraphQL path threw it)
- move resolveWebSocketImpl into config.ts; drop the graphql-ws dependency
- regenerate graphql types (exec operations removed)
- tests: the ws-exec suite becomes the canonical exec suite; e2e reworked
  for WS semantics
- add `mise run e2e` task (loads .env, SANDBOX_E2E=1, --use-system-ca)
- docs: README/AGENTS updated for the WS-only model
