+++ AGENTS.md
@@ -18,7 +18,7 @@ TypeScript SDK for Railway sandbox create/exec/destroy.
 - Public API: `Sandbox` IS the sandbox via static factories — `import { Sandbox } from "railway"`, then `Sandbox.create()` / `Sandbox.connect(id)` / `Sandbox.list()`. The constructor is private; never `new Sandbox(...)`.
 - `token`/`environmentId`/`endpoint` resolve explicit → env (`RAILWAY_API_TOKEN`/`RAILWAY_ENVIRONMENT_ID`/`RAILWAY_GRAPHQL_ENDPOINT`) → default; missing credentials throw `RailwayAuthError`.
 - Keep public SDK scope minimal: `Sandbox.create`, `exec`/`destroy`/`refresh` on the instance, `connect`/`list` statics, `await using` auto-destroy. Files, ports, and richer streaming are future (see `docs/sandbox-sdk-vision.md`).
-- `exec` is durable: the mutation fast-returns; long-running commands stream output over graphql-ws (`sandboxExecOutput` subscription, `resumeToken`/`after`-resumable across the server's ~15m subscription cap; gaps are server-internal, so streaming truncation isn't observable). `exec()` returns an `ExecHandle` (await it for `ExecResult`; `.execId`/`.kill()`); `exec({ execId })` reattaches. The mutation's `waitMs` arg sets the inline-drain window: the SDK sends `waitMs: 0` when `onStdout`/`onStderr` or `timeoutSec` is set (stream from the start), else lets the server default (~25s, short commands inline with no WS); callers can override via `ExecOptions.waitMs`. `timeoutSec` is client-side only (kill + `timedOut: true`; the server arg is gone). Live e2e: `SANDBOX_E2E=1 pnpm vitest run tests/exec.e2e.test.ts` with `.env` creds.
+- `exec` runs over the tcp-proxy `/ws/exec` WebSocket bridge (the only transport): a `shell`-scoped JWT from `generateShellToken` authorizes the socket, which carries separated stdout/stderr and a real exit code. `exec()` returns an `ExecHandle` (await it for `ExecResult`; `.execId`/`.kill()`); `exec({ execId })` reattaches when durable sessions are enabled server-side (VM-assigned id, replays the retained tail then follows live). `kill()`/`timeoutSec` close the socket (with durable on this only detaches, the process keeps running). `timeoutSec` is client-side only (close + `timedOut: true`). The WS endpoint derives from `endpoint` (`backboard.<host>` → `wss://ssh.<host>:2226/ws/exec`); override via `tcpProxyWsEndpoint`/`RAILWAY_TCP_PROXY_WS_ENDPOINT`. Live e2e: `SANDBOX_E2E=1 pnpm vitest run tests/exec.e2e.test.ts` with `.env` creds.
 - Release: pushing a `v*.*.*` tag triggers `release.yml`, which publishes the `package.json` version **at the tagged commit** to npm. Invariant: `package.json` must already equal the tag's version in the commit the tag points to, or `npm publish` fails with "cannot publish over previously published versions".
   - Easiest: run the `Create Release` workflow (Actions → workflow_dispatch, `bump: patch`) — bumps `package.json`, commits to `main`, tags, pushes.
   - Manual: on `main`, `npm version <x.y.z> --no-git-tag-version`, commit the bump, push `main`, then `git tag v<x.y.z>` on that commit and push the tag.
