+++ package.json
@@ -58,8 +58,7 @@
   },
   "dependencies": {
     "@graphql-typed-document-node/core": "^3.2.0",
-    "graphql": "^16.10.0",
-    "graphql-ws": "^6.0.8"
+    "graphql": "^16.10.0"
   },
   "devDependencies": {
     "@graphql-codegen/cli": "^5.0.3",
